Reorganize projects to make it easier to extend it
