+++ catalog/src/main/java/io/kaoto/backend/api/metadata/catalog/CatalogWarmingUpException.java
