+++ kamelet-support/src/main/resources/application.yaml
@@ -1,0 +1,3 @@
+repository:
+  step:
+  viewdefinition:
