+++ kamelet-support/src/test/resources/application.yaml
@@ -1,0 +1,3 @@
+repository:
+  step:
+  viewdefinition:
