+++ pom.xml
@@ -119,6 +119,20 @@
                     </execution>
                 </executions>
             </plugin>
+            <!-- CDI -->
+            <plugin>
+                <groupId>org.jboss.jandex</groupId>
+                <artifactId>jandex-maven-plugin</artifactId>
+                <version>1.2.1</version>
+                <executions>
+                    <execution>
+                        <id>make-index</id>
+                        <goals>
+                            <goal>jandex</goal>
+                        </goals>
+                    </execution>
+                </executions>
+            </plugin>
         </plugins>
     </build>
     <reporting>
@@ -144,6 +158,9 @@
     <modules>
         <module>model</module>
         <module>metadata</module>
+        <module>catalog</module>
+        <module>services-interfaces</module>
+        <module>kamelet-support</module>
         <module>api</module>
     </modules>
     <profiles>
